### 2020-02-01 #2056 @Kevin-Mattheus-Moerman
@egemose thanks for your submission. I recommend adding the software name in the paper title e.g.
`DroneVideoMeasure: Using drones to quantify the length of marine animals`
