### 2020-02-01 #2056 @Kevin-Mattheus-Moerman
@kthyng the authors suggested you as editor. Would you be able to edit this submission? You can likely also invite the authors/reviewers from this submission: https://github.com/openjournals/joss-reviews/issues/1825 to review this work.
If you cannot take this I can take it myself.
