### 2020-02-03 #2056 @kthyng
Hi @egemose! I have a few questions to get started. First, is there any documentation? I didn't see any on a quick scan through your repository.
