### 2020-02-05 #2056 @egemose
Hi @kthyng This is a new submission of https://github.com/openjournals/joss-reviews/issues/1654 where most of the software have been rewritten to work in docker and a browser, to ease the installation and use.
