### 2020-02-05 #2056 @egemose
@kthyng I can see that I forgot to redo the manual of the old repository with the documentation. I will quickly add it.
