+++ paper/paper.bib
@@ -12,3 +12,33 @@
   url = {https://github.com/egemose/DroneVideoMeasure/},
   year = 2020
 }
+
+@article{Dawson2017,
+author = {{Dawson}, S.~M.},
+title = {{Inexpensive Aerial Photogrammetry for Studies of Whales and Large Marine Animals}},
+journal = {Frontier in Marine Science},
+year = {2017},
+volume = {4},
+pages = {366},
+month = nov,
+doi =  {10.3389/fmars.2017.00366}
+}
+
+@article{Burnett2018,
+author = {{Burnett}, J.~D. and {Lemos}, L. and {Barlow}, D. and {Wing}, M~G. and {Chandler}, T. and {Torres}, L.~G.},
+title = {{Estimating morphometric attributes of baleen whales with photogrammetry from small UASs: A case study with blue and gray whales}},
+journal = {Marine Mammal Science},
+year = {2018},
+volume = {35},
+number = {1},
+pages = {108--139},
+doi = {10.1111/mms.12527}
+}
+
+@article{Torres2020,
+author = {Torres et al.}
+title = {MorphoMetriX: a photogrammetric measurement GUI for morphometric analysis of megafauna}
+journal = {Journal of Open Source Software}
+year = {2020}
+doi = {10.21105/joss.01825}
+}
