### 2020-02-06 #2056 @egemose
@kthyng There is now documentation in the "manual" directory.
