+++ paper/paper.bib
@@ -36,9 +36,9 @@ doi = {10.1111/mms.12527}
 }
 
 @article{Torres2020,
-author = {Torres et al.}
-title = {MorphoMetriX: a photogrammetric measurement GUI for morphometric analysis of megafauna}
-journal = {Journal of Open Source Software}
-year = {2020}
+author = {Torres et al.},
+title = {MorphoMetriX: a photogrammetric measurement GUI for morphometric analysis of megafauna},
+journal = {Journal of Open Source Software},
+year = {2020},
 doi = {10.21105/joss.01825}
 }
