### 2020-02-06 #2056 @kthyng
Hi @abhishekbajpayee and @jizecn! You were reviewers on a previous [submission](https://github.com/openjournals/joss-reviews/issues/1654) to JOSS which was paused for some code improvements. It has been resubmitted in this PRE-REVIEW issue. Are you able and willing to review it at this point?
