+++ requirements.txt
@@ -14,4 +14,5 @@ scikit-learn>=0.22
 bokeh==1.0.2
 gunicorn>=19.9.0
 gevent>=1.4.0
-
+celery>=4.4.0
+redis>=3.4.1
