+++ docker-compose.yml
@@ -17,7 +17,6 @@ services:
     build: .
     container_name: dvm_worker
     command: celery worker -A app.celery --loglevel=warn -c 1
-    restart: always
     depends_on:
       - redis
   redis:
