### 2020-02-11 679f81e
fixed extracting images from video
