+++ app_config.py
@@ -9,7 +9,7 @@ class AppConfig:
     # Dropzone settings
     DROPZONE_UPLOAD_MULTIPLE = True
     DROPZONE_MAX_FILE_SIZE = 100000
-    DROPZONE_PARALLEL_UPLOADS = 20
+    DROPZONE_PARALLEL_UPLOADS = 1
     DROPZONE_TIMEOUT = 1800000
     CELERY_BROKER_URL = 'redis://redis:6379/0'
     CELERY_RESULT_BACKEND = 'redis://redis:6379/0'
