### 2020-02-11 6b1bdca
fixed layout of plot
