### 2020-02-11 8a46097
updated windows compose file
