### 2020-02-11 5ba2496
making running containers detached the default
