### 2020-02-13 #2097 @whedon
[REVIEW]: Using drones to quantify the length of marine animals
**Submitting author:** @egemose (<a href="http://orcid.org/0000-0002-6765-8216">Henrik Dyrberg Egemose</a>)
**Repository:** <a href="https://github.com/egemose/DroneVideoMeasure" target ="_blank">https://github.com/egemose/DroneVideoMeasure</a>
**Version:** v1.0.0
