### 2020-02-21 8b0d16b
removed using only docker image because with celery docker compose is needed.
