### 2020-02-21 6e7d4dd
updated manual
