### 2020-02-21 ee6b1b6
updated docker file to use a standard python image
