+++ docker-compose.yml
@@ -15,6 +15,7 @@ services:
       - redis
   worker:
     build: .
+    image: dvm_worker
     container_name: dvm_worker
     command: celery worker -A app.celery --loglevel=warn -c 1
     volumes:
