### 2020-02-21 51e85ca
fixed update time for drones
