### 2020-02-21 0aa1109
validate that a log file is uploaded
