+++ help_functions.py
@@ -14,7 +14,6 @@ def get_last_modified_time(m_type, name=None):
         all_files = glob.glob(os.path.join(base_dir, m_type, name, '*'))
     else:
         all_files = glob.glob(os.path.join(base_dir, m_type, '*'))
-    print(all_files)
     last_m_time = 0
     m_time = 0
     for file in all_files:
