+++ app_config.py
@@ -1,5 +1,8 @@
 import os
 import random
+from celery import Celery
+from flask_dropzone import Dropzone
+from flask_obscure import Obscure
 
 
 class AppConfig:
@@ -13,3 +16,12 @@ class AppConfig:
     DROPZONE_TIMEOUT = 1800000
     CELERY_BROKER_URL = 'redis://redis:6379/0'
     CELERY_RESULT_BACKEND = 'redis://redis:6379/0'
+
+
+base_dir = os.path.abspath('data')
+
+celery = Celery(__name__, broker=AppConfig.CELERY_BROKER_URL, backend=AppConfig.CELERY_RESULT_BACKEND)
+tasks = {}
+
+dropzone = Dropzone()
+obscure = Obscure()
