+++ requirements.txt
@@ -16,3 +16,7 @@ gunicorn>=19.9.0
 gevent>=1.4.0
 celery>=4.4.0
 redis>=3.4.1
+flask-sqlalchemy>=2.4.1
+flask-migrate>=2.5.2
+Flask-Script>=2.0.6
+psycopg2>=2.8.4 
