+++ projects.py
@@ -14,7 +14,7 @@ from werkzeug.utils import secure_filename
 from drone.drone_log_data import drone_log
 from forms import NewProjectForm, EditProjectForm
 from drone.fov import fov
-from app_config import base_dir, celery, tasks
+from app_config import data_dir, celery, tasks
 
 logger = logging.getLogger('app.' + __name__)
 
@@ -38,13 +38,13 @@ def index():
 def make_project_dict():
     logger.debug(f'Finding all projects')
     project_dict = {}
-    project_dir = os.path.join(base_dir, 'projects')
+    project_dir = os.path.join(data_dir, 'projects')
     if not os.path.exists(project_dir):
         os.mkdir(project_dir)
     projects = next(os.walk(project_dir))[1]
     for project_title in projects:
         last_modified_string = ''
-        with open(os.path.join(base_dir, 'projects', project_title, 'description.txt'), 'r') as file:
+        with open(os.path.join(data_dir, 'projects', project_title, 'description.txt'), 'r') as file:
             description = flask.Markup(file.read())
         project = project_tuple(project_title, description, last_modified_string)
         project_dict.update({project_title: project})
@@ -53,7 +53,7 @@ def make_project_dict():
 
 def get_new_project_form(project_dict):
     form = NewProjectForm()
-    drone_dir = os.path.join(base_dir, 'drones')
+    drone_dir = os.path.join(data_dir, 'drones')
     if not os.path.exists(drone_dir):
         os.mkdir(drone_dir)
     drones = next(os.walk(drone_dir))[2]
@@ -67,11 +67,11 @@ def get_new_project_form(project_dict):
         if project_title in project_dict:
             flask.flash('A project with that name already exist!')
         else:
-            os.mkdir(os.path.join(base_dir, 'projects', project_title))
-            with open(os.path.join(base_dir, 'projects', project_title, 'description.txt'), 'w') as file:
+            os.mkdir(os.path.join(data_dir, 'projects', project_title))
+            with open(os.path.join(data_dir, 'projects', project_title, 'description.txt'), 'w') as file:
                 file.write(description)
-            shutil.copyfile(os.path.join(base_dir, 'drones', drone[:-4] + '.cam.npz'), os.path.join(base_dir, 'projects', project_title, 'drone.cam.npz'))
-            log_file = os.path.join(base_dir, 'projects', project_title, 'drone_log.csv')
+            shutil.copyfile(os.path.join(data_dir, 'drones', drone[:-4] + '.cam.npz'), os.path.join(data_dir, 'projects', project_title, 'drone.cam.npz'))
+            log_file = os.path.join(data_dir, 'projects', project_title, 'drone_log.csv')
             form.log_file.data.save(log_file)
             return flask.redirect(flask.url_for('projects.upload', project=project_title)), form
     return None, form
@@ -79,7 +79,7 @@ def get_new_project_form(project_dict):
 
 def get_edit_project_form(project_dict):
     form = EditProjectForm()
-    drones = next(os.walk(os.path.join(base_dir, 'drones')))[2]
+    drones = next(os.walk(os.path.join(data_dir, 'drones')))[2]
     drones = [x for x in drones if x.endswith('.txt')]
     form.edit_drone.choices = [(d, d[:-4]) for d in drones]
     if form.validate_on_submit():
@@ -92,19 +92,19 @@ def get_edit_project_form(project_dict):
             flask.flash('A project with that name already exist!')
         else:
             project_dict.pop(project_before, None)
-            shutil.copyfile(os.path.join(base_dir, 'drones', drone[:-4] + '.cam.npz'), os.path.join(base_dir, 'projects', project_title, 'drone.cam.npz'))
[…]
