+++ drone/drones.py
@@ -1,91 +1,77 @@
 import os
-import csv
 import re
 import shutil
 import numpy as np
-from collections import namedtuple
 import flask
 import json
 import logging
+from datetime import datetime
 from werkzeug.utils import secure_filename
 from forms import NewDroneForm, EditDroneForm
-from app_config import base_dir
+from app_config import data_dir, db
 from calibration.calibration import CalibrateCamera
 
 logger = logging.getLogger('app.' + __name__)
-
-
 drones_view = flask.Blueprint('drones', __name__)
-drone_tuple = namedtuple('drone', ['title', 'camera_settings', 'calibrated', 'last_updated'])
+
+
+class Drone(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(), unique=True, nullable=False)
+    description = db.Column(db.Text())
+    calibration = db.Column(db.PickleType())
+    last_update = db.Column(db.DateTime())
+
+    def __repr__(self):
+        return f'<Drone {self.name}>'
 
 
 @drones_view.route('/drones', methods=['GET', 'POST'])
 def index():
-    drone_dict = make_drone_dict()
-    drone_form = get_drone_form(drone_dict)
-    edit_drone_form = get_edit_drone_form(drone_dict)
+    drone_form = get_drone_form()
+    edit_drone_form = get_edit_drone_form()
+    drones = Drone.query.all()
     logger.debug(f'Render drone index')
-    return flask.render_template('drones/index.html', drones=drone_dict, new_drone_form=drone_form, edit_drone_form=edit_drone_form)
+    return flask.render_template('drones/index.html', drones=drones, new_drone_form=drone_form, edit_drone_form=edit_drone_form)
 
 
-def get_drone_form(drone_dict):
+def get_drone_form():
+    drones = Drone.query.all()
     form = NewDroneForm()
     if form.validate_on_submit():
-        drone_title = form.name.data
+        drone_name = form.name.data
         camera_settings = form.camera_settings.data
-        if drone_title in drone_dict:
-            flask.flash('A project with that name already exist!')
+        if drone_name in [x.name for x in drones]:
+            flask.flash('A drone with that name already exist!')
         else:
-            with open(os.path.join(base_dir, 'drones',  drone_title + '.txt'), 'w') as file:
-                writer = csv.writer(file)
-                header = ('title', 'camera_settings')
-                writer.writerow(header)
-                writer.writerow((drone_title, camera_settings))
-            drone = drone_tuple(drone_title, camera_settings, False, '')
-            drone_dict.update({drone.title: drone})
+            drone = Drone(name=drone_name, description=camera_settings, last_update=datetime.now())
+            db.session.add(drone)
+            db.session.commit()
     return form
 
 
-def get_edit_drone_form(drone_dict):
+def get_edit_drone_form():
+    drones = Drone.query.all()
     form = EditDroneForm()
     if form.validate_on_submit():
-        project_before = form.edit_drone_before.data
-        drone_title = form.edit_name.data
-        camera_settings = form.edit_camera_settings.data
-        if drone_title in drone_dict and not drone_title == project_before:
-            flask.flash('A project with that name already exist!')
+        drone_id = form.edit_drone_id.data
+        drone_name_before = form.edit_drone_before.data
+        new_drone_name = form.edit_name.data
+        new_camera_settings = form.edit_camera_settings.data
+        drone = Drone.query.get_or_404(drone_id)
+        if drone.name in [x.name for x in drones] and not drone.name == drone_name_before:
+            flask.flash('A drone with that name already exist!')
         else:
[…]
