+++ migrations/versions/76afd823e8bb_added_drone.py
@@ -1,0 +1,36 @@
+"""added Drone
+
+Revision ID: 76afd823e8bb
+Revises: 
+Create Date: 2020-02-25 13:35:21.837709
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '76afd823e8bb'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('drone',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('cal_file', sa.String(), nullable=True),
+    sa.Column('last_update', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('drone')
+    # ### end Alembic commands ###
