+++ migrations/versions/2e034d73923e_.py
@@ -1,0 +1,36 @@
+"""empty message
+
+Revision ID: 2e034d73923e
+Revises: 8772d6966f4f
+Create Date: 2020-03-02 13:54:16.091201
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '2e034d73923e'
+down_revision = '8772d6966f4f'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('task', sa.Column('drone_id', sa.Integer(), nullable=True))
+    op.alter_column('task', 'video_id',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.create_foreign_key(None, 'task', 'drone', ['drone_id'], ['id'])
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, 'task', type_='foreignkey')
+    op.alter_column('task', 'video_id',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    op.drop_column('task', 'drone_id')
+    # ### end Alembic commands ###
