### 2020-03-12 #2097 @abhishekbajpayee
@kthyng apologies for the delay.

I installed things and launched the application using provided instructions. Using calibration video in the sample dataset, I was able to calibrate. I did notice a few things that might be worth addressing here:
1) The UI to pick a calibration video file only shows mp4 video files but the sample file is an mov file. Dragging and dropping worked for me but seeing only mp4 in the file selection UI was a bit confusing.
2) It might be useful to either show progress or logs or some sort while calibration is going on. User has no idea how long to wait.
3) Ideally there should be some way to handle incorrect calibration videos i.e. videos that do not contain a checkerboard. Initially when I just saw the mp4 files in the file selector, I picked the demo.mp4 files instead but seems like things froze. So I had to restart everything.

Once beyond calibration, I was able to create a project and upload a video. Comments for this step:
1) Here I wasn't sure which video or videos in the sample data correspond to which csv log file.
2) Not knowing the correspondence, I tried all combinations but got stuck after the upload. So the upload completed but then going to the video gallery didn't show any videos. Not sure how to proceed from here.

@egemose can you advise on how I should proceed with testing after calibration?
