### 2020-03-12 #2097 @jizecn
Hi, I also tried the software following the instructions provided.

Docker to facilitate the installation and testing works fine for me. However, I need to use "sudo" to start the software. Maybe it is only for me, but might be worth mentioning in the document (ps. I use Linux Ubuntu 16.04).

A few comments/suggestions:
1. The calibration process looks similar to the annotation part. I by mistake uploaded a wrong video for calibration, and that took ages without notifying any error.
2. Similarly, when creating a project, I also uploaded a wrong file for the log file (which I initially thought it was for the video). However, the system did not report any error either until at a later stage. It might be useful to use tooltips providing instructions about at least file types expected to be uploaded and the content of the file.
3. The UI buttons: can we change the appearance to make it look clearer when toggled? The three buttons for changing colours, line width, and removing annotations are useful, but not very intuitive in my opinion. Can we provide some tooltips too when the mouse pointer hovers around the buttons for these buttons?

As mentioned by the other reviewer (@abhishekbajpayee ), it is not clear how to identify corresponding log and video files. I was only able to do it following the demo video. It is worth providing a simple tool (e.g. a python / bash script) that can rename two paired files after downloading from the drone to the same one (as a pre-processing stage), so that user only needs to upload two files once.

Overall, I can confirm the software works as described (overall good documentation), except the user interface can be improved for better usability. I hope it helps. Thanks.
