+++ projects/video_gallery.py
@@ -1,0 +1,253 @@
+
+import json
+import os
+import random
+import re
+import subprocess
+import logging
+import flask
+import ffmpeg
+import numpy as np
+import csv
+from drone.fov import fov
+from drone.drone_log_data import drone_log
+from app_config import data_dir, get_random_filename, celery, Task, Project, Video, Drone, db
+
+
+logger = logging.getLogger('app.' + __name__)
+video_gallery_view = flask.Blueprint('video_gallery', __name__)
+
+
+@video_gallery_view.route('/videos/<project_id>/upload', methods=['GET', 'POST'])
+def upload(project_id):
+    if flask.request.method == 'POST':
+        for key, file in flask.request.files.items():
+            if key.startswith('file'):
+                video_file = os.path.join(data_dir, get_random_filename(file.filename.rsplit('.', 1)[0] + '.mp4'))
+                video_mime_type = re.compile('video/*')
+                if video_mime_type.match(file.mimetype):
+                    temp_filename = get_random_filename(file.filename)
+                    logger.debug(f'Uploading file: {video_file}')
+                    temp_file = os.path.join(data_dir, temp_filename)
+                    file.save(temp_file)
+                    logger.debug(f'Upload done for file: {video_file}')
+                    logger.debug(f'Calling celery to convert file: {temp_filename} and save as {video_file}')
+                    video = Video(file=video_file, name=file.filename, project_id=project_id, image=video_file + '.jpg')
+                    db.session.add(video)
+                    db.session.commit()
+                    task = convert_after_upload_task.apply_async(args=(temp_file, video.file))
+                    task_db = Task(task_id=task.id, function='convert_after_upload_task', video_id=video.id)
+                    db.session.add(task_db)
+                    db.session.commit()
+                    return flask.jsonify({}), 202
+    logger.debug(f'Render upload after a GET request')
+    project = Project.query.get_or_404(project_id)
+    return flask.render_template('video_gallery/upload.html', project_id=project.id, project_name=project.name)
+
+
+def get_video_data(video_file, location_video_file=None):
+    logger.debug(f'Reading video data for video {video_file}')
+    ffprobe_res = ffmpeg.probe(video_file, cmd='ffprobe')
+    video_duration = float(ffprobe_res['format']['duration'])
+    video_nb_frames = int(ffprobe_res['streams'][0]['nb_frames'])
+    video_size = (int(ffprobe_res['streams'][0]['width']), int(ffprobe_res['streams'][0]['height']))
+    try:
+        location_string = ffprobe_res['format']['tags']['location']
+    except KeyError:
+        if location_video_file:
+            ffprobe_res = ffmpeg.probe(location_video_file, cmd='ffprobe')
+            location_string = ffprobe_res['format']['tags']['location']
+        else:
+            location_string = None
+    video_pos = (None, None)
+    if location_string:
+        match = re.match(r'([-+]\d+.\d+)([-+]\d+.\d+)([-+]\d+.\d+)', location_string)
+        if match:
+            video_pos = (float(match.group(1)), float(match.group(2)))
+    return video_duration, video_nb_frames, video_size, video_pos
+
+
+@celery.task(bind=True)
+def convert_after_upload_task(self, temp_path, video_path):
+    self.update_state(state='PROCESSING')
[…]
