+++ drone/drones.py
@@ -7,7 +7,7 @@ import logging
 from datetime import datetime
 from werkzeug.utils import secure_filename
 from forms import NewDroneForm, EditDroneForm
-from app_config import data_dir, celery, Drone, Task, db
+from app_config import data_dir, celery, Drone, Task, db, TaskFailure
 from calibration.calibration import CalibrateCamera
 
 logger = logging.getLogger('app.' + __name__)
@@ -15,15 +15,17 @@ drones_view = flask.Blueprint('drones', __name__)
 
 
 @drones_view.route('/drones', methods=['GET', 'POST'])
-def index():
-    drone_form = get_drone_form()
+def drones():
+    res, drone_form = get_new_drone_form()
+    if res:
+        return res
     edit_drone_form = get_edit_drone_form()
     drones = Drone.query.all()
     logger.debug(f'Render drone index')
-    return flask.render_template('drones/index.html', drones=drones, new_drone_form=drone_form, edit_drone_form=edit_drone_form)
+    return flask.render_template('drones/drones.html', drones=drones, new_drone_form=drone_form, edit_drone_form=edit_drone_form)
 
 
-def get_drone_form():
+def get_new_drone_form():
     drones = Drone.query.all()
     form = NewDroneForm()
     if form.validate_on_submit():
@@ -32,10 +34,11 @@ def get_drone_form():
         if drone_name in [x.name for x in drones]:
             flask.flash('A drone with that name already exist!')
         else:
-            drone = Drone(name=drone_name, description=camera_settings, last_update=datetime.now())
+            drone = Drone(name=drone_name, description=camera_settings)
             db.session.add(drone)
             db.session.commit()
-    return form
+            return flask.redirect(flask.url_for('drones.add_calibration', drone_id=drone.id)), form
+    return None, form
 
 
 def get_edit_drone_form():
@@ -86,34 +89,39 @@ def do_calibration(drone_id):
     logger.debug(f'do_calibration is called for drone {drone_id}')
     drone = Drone.query.get_or_404(drone_id)
     drone.calibration = None
+    drone.task_error = None
     db.session.commit()
-    task = concat_videos_task.apply_async(args=(drone.id,))
-    task_db = Task(task_id=task.id, function='concat_videos_task', drone_id=drone.id)
+    task = calibration_task.apply_async(args=(drone.id,))
+    task_db = Task(task_id=task.id, function='calibration_task', drone_id=drone.id)
     db.session.add(task_db)
     db.session.commit()
     return ''
 
 
 @celery.task(bind=True)
-def concat_videos_task(self, drone_id):
+def calibration_task(self, drone_id):
     self.update_state(state='PROCESSING')
     in_folder = os.path.join(data_dir, 'calibration')
     calibrate_cam = CalibrateCamera()
     try:
         result = calibrate_cam(in_folder)
         if result == -1:
-            self.update_state(state='Error', meta={'type': 'danger', 'message': 'Error finding checkerboard during calibration! Please try again with new images or new video by reloading the page.'})
-            return False
+            shutil.rmtree(in_folder)
+            os.mkdir(in_folder)
+            raise TaskFailure('Error: Could not find any checkerboards in the images/video. Try with a new video.')
         if not result:
-            self.update_state(state='Error', meta={'type': 'danger', 'message': 'Error no video or images found! Please try again by reloading the page.'})
-            return False
+            shutil.rmtree(in_folder)
+            os.mkdir(in_folder)
+            raise TaskFailure('Error: No video or images found. Please try again.')
         drone_db = Drone.query.get(drone_id)
         drone_db.calibration = result
[…]
