### 2020-08-19 bd664b6
do not stop if failed
