+++ Dockerfile
@@ -1,5 +1,5 @@
 FROM python:3.6 as base
-RUN curl -sL https://deb.nodesource.com/setup_13.x | bash - \
+RUN curl -sL https://deb.nodesource.com/setup_12.x | bash - \
 && apt-get update && apt-get install -y \
     ffmpeg \
     nodejs \
