+++ README.md
@@ -4,22 +4,23 @@ A tool to measure and track things on a planar surface.
 Developed to measure and track marine mammals in the surface of the ocean.
 
 * [Getting Started](#getting-started)
-  - [Using a Docker-hub image](#using-a-docker-hub-image-the-easy-way)
-  - [Manual install](#manual-install)
-  - [Usage](#usage)
-  - [Author](#author)
-  - [License](#license)
+  * [Using a Docker-hub image](#using-a-docker-hub-image-the-easy-way)
+  * [Manual install](#manual-install)
+  * [Usage](#usage)
+  * [Author](#author)
+  * [License](#license)
 
 ## Getting Started
 
 ### Manual install
 
 * Install the following applications:
-  - [Git](https://git-scm.com/downloads)
-  - [Docker](https://www.docker.com/)
-  - [Docker-compose](https://docs.docker.com/compose/install/) (only on Linux)
+  * [Git](https://git-scm.com/downloads)
+  * [Docker](https://www.docker.com/)
+  * [Docker-compose](https://docs.docker.com/compose/install/) (only on Linux)
 
 * From Powershell (windows) or command line (Mac / Linux), type:
+
 ```bash
 git clone https://github.com/egemose/DroneVideoMeasure.git
 cd DroneVideoMeasure
@@ -27,30 +28,36 @@ cd DroneVideoMeasure
 ```
 
 * Open a Web Browser at `http://localhost:5000`
-  - If Drone Video Measure appears all is set and up and running.
+  * If Drone Video Measure appears all is set and up and running.
 
 * To stop Drone Video Measure, type:
+
 ```bash
 ./dvm.sh stop
 ```
 
 * To rebuild Drone Video Measure, type:
+
 ```bash
 ./dvm.sh rebuild
 ```
 
 * To update Drone Video Measure, type:
+
 ```bash
 ./dvm.sh update
 ```
 
 * To uninstall Drone Video Measure, type:
+
 ```bash
 ./dvm.sh remove
 ```
+
 And delete the DroneVideoMeasure directory.
 
 * On Windows data is stored in a Docker volume and will persist even when DVM is closed. To remove the data, type:
+
 ```bash
 docker volume rm appmedia
 ```
@@ -67,7 +74,7 @@ Demo data is available at [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.36
 
 * DJI
 
-Currently only flight logs from DJI drones are suppored. If you what support for other drones please create a issue with a flight log attached.
+Currently only flight logs from DJI drones are supported. If you what support for other drones please create a issue with a flight log attached.
 
 ## Author
 
