+++ migrations/versions/8aa202e6821d_added_error_on_project.py
@@ -1,0 +1,34 @@
+"""added error on project
+
+Revision ID: 8aa202e6821d
+Revises: 4d9e0a758bf7
+Create Date: 2020-05-07 13:27:03.016882
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '8aa202e6821d'
+down_revision = '4d9e0a758bf7'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('project', sa.Column('log_error', sa.String(), nullable=True))
+    op.alter_column('project', 'log_file',
+               existing_type=sa.VARCHAR(),
+               nullable=True)
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('project', 'log_file',
+               existing_type=sa.VARCHAR(),
+               nullable=False)
+    op.drop_column('project', 'log_error')
+    # ### end Alembic commands ###
