+++ migrations/versions/4d9e0a758bf7_added_task_error.py
@@ -1,0 +1,30 @@
+"""added task error
+
+Revision ID: 4d9e0a758bf7
+Revises: f43cf70961bb
+Create Date: 2020-05-05 20:04:31.679365
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '4d9e0a758bf7'
+down_revision = 'f43cf70961bb'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('drone', sa.Column('task_error', sa.String(), nullable=True))
+    op.add_column('video', sa.Column('task_error', sa.String(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('video', 'task_error')
+    op.drop_column('drone', 'task_error')
+    # ### end Alembic commands ###
